fix a bug in pic_false_value

+++ include/picrin/value.h
@@ -43,7 +43,7 @@ typedef struct {
 
 #define pic_ptr(v) ((void *)((long long)0xffffffffffff & (long long)(v).u.data))
 #define pic_vtype(v) (((v).u.type_ & 0xf0000)>>16)
-#define pic_init_value(v,vtype) (((v).u.type_ = ((unsigned int)0xfff00000|((vtype)<<16))), (v).u.i = 0)
+#define pic_init_value(v,vtype) (((v).u.type_ = ((unsigned int)0xfff00000|(unsigned int)((vtype)<<16))), (v).u.i = 0)
 
 #else
